Cleanup and remove changes
