+++ netket/_src/ngd/sr.py
@@ -6,7 +6,6 @@ import jax.numpy as jnp
 from jax.sharding import PositionalSharding
 
 from netket import jax as nkjax
-from netket.utils import mpi
 from netket.utils.types import Array
 
 from netket._src.ngd.kwargs import ensure_accepts_kwargs
@@ -39,35 +38,23 @@ def _compute_sr_update(
         raise ValueError("Not implemented")
 
     # (np, #ns) x (#ns) -> (np) - where the sum over #ns is done automatically
-    # in sharding, while under MPI we need to do it manually with an allreduce_sum.
-    F, token = mpi.mpi_allreduce_sum_jax(O_L.T @ dv, token=None)
+    F = O_L.T @ dv
 
     # This does the contraction (np, #ns) x (#ns, np) -> (np, np).
-    # When using sharding the sum over #ns is done automatically.
-    # When using MPI we need to do it manually with an allreduce_sum.
-    matrix, token = mpi.mpi_reduce_sum_jax(O_L.T @ O_L, root=0, token=token)
+    matrix = O_L.T @ O_L
     matrix_side = matrix.shape[-1]  # * it can be ns or 2*ns, depending on mode
 
-    if mpi.rank == 0:
-        shifted_matrix = jax.lax.add(
-            matrix, diag_shift * jnp.eye(matrix_side, dtype=matrix.dtype)
-        )
-        updates = solver_fn(shifted_matrix, F, dv=dv)
+    shifted_matrix = jax.lax.add(
+        matrix, diag_shift * jnp.eye(matrix_side, dtype=matrix.dtype)
+    )
+    updates = solver_fn(shifted_matrix, F, dv=dv)
 
-        # Some solvers return a tuple, some others do not.
-        if isinstance(updates, tuple):
-            updates, info = updates
-        else:
+    # Some solvers return a tuple, some others do not.
+    if isinstance(updates, tuple):
+        updates, info = updates
+        if info is None:
             info = {}
-
-        updates = updates.reshape(mpi.n_nodes, -1)
-        updates, token = mpi.mpi_scatter_jax(updates, root=0, token=token)
     else:
-        updates = jnp.zeros((int(matrix_side / mpi.n_nodes),), dtype=jnp.float64)
-        updates, token = mpi.mpi_scatter_jax(updates, root=0, token=token)
-        info = None
-
-    if info is None:
         info = {}
 
     # If complex mode and we have complex parameters, we need
