+++ pyproject.toml
@@ -160,7 +160,7 @@ exclude = ["Examples", "docs/conf.py", "*.ipynb"]
 [tool.ruff.lint]
 select = ["E", "F", "W", "D410", "UP"]
 fixable = ["E", "F", "W", "D", "UP"]
-ignore = ["E501", "E731", "E721", "E741", "UP007", "UP035", "UP038"]
+ignore = ["E501", "E731", "E721", "E701", "E741", "UP007", "UP035", "UP038"]
 #ignore = ["E266"]
 
 [tool.ruff.lint.per-file-ignores]
